+++ sharp_injector/Helpers/PredictionWindowHelper.cs
@@ -18,26 +18,26 @@ namespace sharp_injector.Helpers {
     internal static class PredictionWindowHelper {
 
         private enum PredictionType {
-            SpeechModelPredictions = 0,
-            DictionaryPredictions
+            PrimaryPredictions = 0,
+            SecondaryPredictions
         }
 
         private static string GetPredictionStackPanelName(PredictionType pt) {
             switch (pt) {
-                case PredictionType.SpeechModelPredictions:
-                    return "SpeechModelPredictions";
-                case PredictionType.DictionaryPredictions:
-                    return "DictionaryPredictions";
+                case PredictionType.PrimaryPredictions:
+                    return "PrimaryPredictions";
+                case PredictionType.SecondaryPredictions:
+                    return "SecondaryPredictions";
                 default:
                     return "";
             }
         }
 
-        
-            
-               
 
-        private static PredictionType curPredictionType = PredictionType.SpeechModelPredictions;
+
+
+
+        private static PredictionType curPredictionType = PredictionType.PrimaryPredictions;
 
         public static void IncrementSelection(Window predictionWindow) {
             predictionWindow.Dispatcher.Invoke(new Action(() => {
@@ -45,8 +45,8 @@ namespace sharp_injector.Helpers {
                     //_selectedPrediction
                     var pwType = predictionWindow.GetType();
                     var selectedPrediction = ((System.Windows.Controls.DockPanel)pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow));
-                    if (selectedPrediction is null || !DictionaryPredictionsVisible) {
-                        curPredictionType = PredictionType.SpeechModelPredictions;
+                    if (selectedPrediction is null || !DictionaryPredictionsVisible(predictionWindow)) {
+                        curPredictionType = PredictionType.PrimaryPredictions;
                     }
                     var SpeechModelPredictions = (StackPanel)pwType.GetField(GetPredictionStackPanelName(curPredictionType), BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                     int next = 0;
@@ -60,14 +60,11 @@ namespace sharp_injector.Helpers {
                         //var predictionType = selectedPrediction.Tag == "SpeechModelPrediction" ? 0 : 1;
                         var curSelected = pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                         pwType.GetMethod("RestorePredictionColor", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { curSelected });
-
-
                     }
 
-                    Terminal.Print($"Next: {next}\n");
+
                     pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).SetValue(predictionWindow, (DockPanel)((Decorator)SpeechModelPredictions.Children[next]).Child);
                     pwType.GetMethod("HighlightPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { (DockPanel)((Decorator)SpeechModelPredictions.Children[next]).Child });
-                    Terminal.Print($"HighlightPrediction Next: {next}\n");
 
                     // Read preidction if enabled
                     if (ReadPredictions) {
@@ -91,15 +88,14 @@ namespace sharp_injector.Helpers {
                     //_selectedPrediction
                     var pwType = predictionWindow.GetType();
                     var selectedPrediction = ((System.Windows.Controls.DockPanel)pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow));
-                    if (selectedPrediction is null || !DictionaryPredictionsVisible) {
-                        curPredictionType = PredictionType.SpeechModelPredictions;
+                    if (selectedPrediction is null || !DictionaryPredictionsVisible(predictionWindow)) {
+                        curPredictionType = PredictionType.PrimaryPredictions;
                     }
                     var SpeechModelPredictions = (StackPanel)pwType.GetField(GetPredictionStackPanelName(curPredictionType), BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                     var temp = new List<Decorator>(SpeechModelPredictions.Children.OfType<Decorator>());
                     int next = temp.Where(x => x.Visibility == Visibility.Visible).Count() - 1;
                     if (!(selectedPrediction is null)) {
                         next = temp.FindIndex(x => (DockPanel)x.Child == selectedPrediction) - 1;
-                        Terminal.Print($"Before Next: {next}\n");
                         if (next < 0) {
                             next = temp.Where(x => x.Visibility == Visibility.Visible).Count() - 1;
                             DecrementPage(predictionWindow);
@@ -111,10 +107,8 @@ namespace sharp_injector.Helpers {
 
                     }
 
-                    Terminal.Print($"Next: {next}\n");
                     pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).SetValue(predictionWindow, (DockPanel)((Decorator)SpeechModelPredictions.Children[next]).Child);
                     pwType.GetMethod("HighlightPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { (DockPanel)((Decorator)SpeechModelPredictions.Children[next]).Child });
-                    Terminal.Print($"HighlightPrediction Next: {next}\n");
 
                     // Read preidction if enabled
                     if (ReadPredictions) {
@@ -133,19 +127,19 @@ namespace sharp_injector.Helpers {
         }
 
         public static bool IncrementPage(Window predictionWindow, bool wrap = false) {
-            if (!DictionaryPredictionsVisible) {
-                curPredictionType = PredictionType.SpeechModelPredictions;
+            if (!DictionaryPredictionsVisible(predictionWindow)) {
+                curPredictionType = PredictionType.PrimaryPredictions;
             }
             var pwType = predictionWindow.GetType();
             var _currentPredictions = pwType.GetField("_currentPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
-            var _currentSpeechModelPage = (int)pwType.GetField("_currentSpeechModelPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
-            var _currentDictionaryPage = (int)pwType.GetField("_currentDictionaryPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+            var _currentSpeechModelPage = (int)pwType.GetField("_currentPrimaryPredictionsPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+            var _currentDictionaryPage = (int)pwType.GetField("_currentSecondaryPredictionsPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
             // Set NavigatingPredictions to false
             var _service = pwType.GetField("_service", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
             var _serviceType = _service.GetType();
             _serviceType.GetField("NavigatingPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).SetValue(_service, false);
-            if ((curPredictionType == PredictionType.SpeechModelPredictions)) {
-                var _currentSpeechModelMaxPages = (int)pwType.GetField("_currentSpeechModelMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+            if ((curPredictionType == PredictionType.PrimaryPredictions)) {
+                var _currentSpeechModelMaxPages = (int)pwType.GetField("_currentPrimaryPredictionsMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                 if (_currentSpeechModelPage >= _currentSpeechModelMaxPages) {
                     if (wrap) {
                         pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, 1, _currentDictionaryPage });
@@ -154,60 +148,60 @@ namespace sharp_injector.Helpers {
                 }
 
                 pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelPage + 1, _currentDictionaryPage });
-                
+
                 return true;
             }
-            var _currentDictionaryMaxPages = (int)pwType.GetField("_currentDictionaryMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+            var _currentDictionaryMaxPages = (int)pwType.GetField("_currentSecondaryPredictionsPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
             if (_currentDictionaryPage >= _currentDictionaryMaxPages) {
                 if (wrap) {
                     pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelPage, 1 });
                 }
-                
+
                 return false;
             }
 
             pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelPage, _currentDictionaryPage + 1 });
-            
+
             return true;
         }
 
         public static bool DecrementPage(Window predictionWindow, bool wrap = false) {
-            if (!DictionaryPredictionsVisible) {
-                curPredictionType = PredictionType.SpeechModelPredictions;
+            if (!DictionaryPredictionsVisible(predictionWindow)) {
+                curPredictionType = PredictionType.PrimaryPredictions;
             }
             var pwType = predictionWindow.GetType();
             var _currentPredictions = pwType.GetField("_currentPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
-            var _currentSpeechModelPage = (int)pwType.GetField("_currentSpeechModelPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
-            var _currentDictionaryPage = (int)pwType.GetField("_currentDictionaryPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
-            
+            var _currentSpeechModelPage = (int)pwType.GetField("_currentPrimaryPredictionsPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+            var _currentDictionaryPage = (int)pwType.GetField("_currentSecondaryPredictionsPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+
             // Set NavigatingPredictions to false
             var _service = pwType.GetField("_service", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
             var _serviceType = _service.GetType();
             _serviceType.GetField("NavigatingPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).SetValue(_service, false);
-            if ((curPredictionType == PredictionType.SpeechModelPredictions)) {
+            if ((curPredictionType == PredictionType.PrimaryPredictions)) {
                 if (_currentSpeechModelPage <= 1) {
                     if (wrap) {
-                        var _currentSpeechModelMaxPages = (int)pwType.GetField("_currentSpeechModelMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+                        var _currentSpeechModelMaxPages = (int)pwType.GetField("_currentPrimaryPredictionsMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                         pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelMaxPages, _currentDictionaryPage });
                     }
-                    
+
                     return false;
                 }
 
                 pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelPage - 1, _currentDictionaryPage });
-                
+
                 return true;
             }
             if (_currentDictionaryPage <= 1) {
                 if (wrap) {
-                    var _currentDictionaryMaxPages = (int)pwType.GetField("_currentDictionaryMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+                    var _currentDictionaryMaxPages = (int)pwType.GetField("_currentSecondaryPredictionsPage", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                     pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelPage, _currentDictionaryMaxPages });
                 }
-                
+
                 return false;
             }
             pwType.GetMethod("ShowPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(predictionWindow, new object[] { _currentPredictions, _currentSpeechModelPage, _currentDictionaryPage - 1 });
-            
+
             return true;
         }
 
@@ -216,13 +210,13 @@ namespace sharp_injector.Helpers {
             var pwType = predictionWindow.GetType();
             var selectedPrediction = ((System.Windows.Controls.DockPanel)pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow));
             if (selectedPrediction is null) {
-                curPredictionType = curPredictionType == PredictionType.DictionaryPredictions ? PredictionType.SpeechModelPredictions : PredictionType.DictionaryPredictions;
+                curPredictionType = curPredictionType == PredictionType.SecondaryPredictions ? PredictionType.PrimaryPredictions : PredictionType.SecondaryPredictions;
                 return;
             }
             var curPredictions = (StackPanel)pwType.GetField(GetPredictionStackPanelName(curPredictionType), BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
             var curPredList = new List<Decorator>(curPredictions.Children.OfType<Decorator>());
             var curIdx = curPredList.FindIndex(x => (DockPanel)x.Child == selectedPrediction);
-            curPredictionType = curPredictionType == PredictionType.DictionaryPredictions ? PredictionType.SpeechModelPredictions : PredictionType.DictionaryPredictions;
+            curPredictionType = curPredictionType == PredictionType.SecondaryPredictions ? PredictionType.PrimaryPredictions : PredictionType.SecondaryPredictions;
             var nextPredictions = (StackPanel)pwType.GetField(GetPredictionStackPanelName(curPredictionType), BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
             var nextPredList = new List<Decorator>(nextPredictions.Children.OfType<Decorator>());
 
@@ -253,12 +247,14 @@ namespace sharp_injector.Helpers {
             _serviceType.GetField("NavigatingPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).SetValue(_service, true);
         }
 
-        public static bool DictionaryPredictionsVisible {
-            get {
-                var Configuration = Type.GetType("AppWriter.ConfigurationManager,AppWriter").GetProperty("Configuration", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(null);
-                var ConfigurationType = Configuration.GetType();
-                return (bool)ConfigurationType.GetProperty("DictionaryPredictionsVisible", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(Configuration);
-            }
+        public static bool DictionaryPredictionsVisible(Window predictionWindow) {
+            var pwType = predictionWindow.GetType();
+            var AppWriterService = pwType.GetField("_service", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+            var AppWriterServiceType = AppWriterService.GetType();
+            var Profile = AppWriterServiceType.GetProperty("Profile", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(AppWriterService);
+            var ProfileType = Profile.GetType();
+
+            return (bool)ProfileType.GetField("GlossaryPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(Profile);
         }
 
         //
@@ -269,10 +265,14 @@ namespace sharp_injector.Helpers {
                 return (bool)ConfigurationType.GetProperty("ReadPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(Configuration);
             }
         }
+        public static int SecondaryPredictionsCount(Window predictionWindow) {
 
+            var pwType = predictionWindow.GetType();
+            return (int)pwType.GetField("_currentSecondaryPredictionsMaxPages", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
+        }
         public static void TogglePredictioTypeOrIncrement(Window predictionWindow) {
             predictionWindow.Dispatcher.Invoke(new Action(() => {
-                if (DictionaryPredictionsVisible) {
+                if (DictionaryPredictionsVisible(predictionWindow) && SecondaryPredictionsCount(predictionWindow) > 0) {
                     TogglePredictionType(predictionWindow);
                 } else {
                     IncrementPage(predictionWindow, true);
@@ -281,16 +281,15 @@ namespace sharp_injector.Helpers {
         }
         public static void TogglePredictioTypeOrDecrement(Window predictionWindow) {
             predictionWindow.Dispatcher.Invoke(new Action(() => {
-                if (DictionaryPredictionsVisible) {
+                if (DictionaryPredictionsVisible(predictionWindow) && SecondaryPredictionsCount(predictionWindow) > 0) {
                     TogglePredictionType(predictionWindow);
                 } else {
                     DecrementPage(predictionWindow, true);
                 }
             }));
         }
-        
-        public static bool InsertSelectedPrediction(Window predictionWindow) {
 
+        public static bool InsertSelectedPrediction(Window predictionWindow) {
             var pwType = predictionWindow.GetType();
             var selectedPrediction = ((System.Windows.Controls.DockPanel)pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow));
             if (selectedPrediction is null) {
@@ -306,7 +305,9 @@ namespace sharp_injector.Helpers {
                 var _service = pwType.GetField("_service", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                 var _serviceType = _service.GetType();
                 var _currentPredictions = pwType.GetField("_currentPredictions", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
-                _serviceType.GetMethod("InsertPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(_service, new object[] { _currentPredictions, curPredictionType, index });
+                var currentPredictionsType = _currentPredictions.GetType();
+                var PredictionResult = currentPredictionsType.GetProperty("PredictionResult", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(_currentPredictions);
+                _serviceType.GetMethod("InsertPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).Invoke(_service, new object[] { PredictionResult, curPredictionType, index });
                 predictionWindow.Visibility = Visibility.Collapsed;
 
                 // Set NavigatingPredictions to false
@@ -317,10 +318,10 @@ namespace sharp_injector.Helpers {
 
 
         public static void SelectPredictionIndex(Window predictionWindow, int index) {
-            if(predictionWindow is null || !predictionWindow.IsVisible) {
+            if (predictionWindow is null || !predictionWindow.IsVisible) {
                 return;
             }
-            curPredictionType = PredictionType.SpeechModelPredictions;
+            curPredictionType = PredictionType.PrimaryPredictions;
             var pwType = predictionWindow.GetType();
             var selectedPrediction = ((System.Windows.Controls.DockPanel)pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow));
             if (!(selectedPrediction is null)) {
@@ -344,7 +345,7 @@ namespace sharp_injector.Helpers {
         }
 
         public static void DeselectPrediction(Window predictionWindow) {
-            curPredictionType = PredictionType.SpeechModelPredictions;
+            curPredictionType = PredictionType.PrimaryPredictions;
             var pwType = predictionWindow.GetType();
             var selectedPrediction = ((System.Windows.Controls.DockPanel)pwType.GetField("_selectedPrediction", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow));
             if (!(selectedPrediction is null)) {
@@ -364,7 +365,7 @@ namespace sharp_injector.Helpers {
         private static Queue<UpdateShortcutTextDelegate> toUpdate = new Queue<UpdateShortcutTextDelegate>();
 
         private static void Postfix_Prediction_Window_Loaded(ref Window __instance) {
-            while(toUpdate.Any()) {
+            while (toUpdate.Any()) {
                 (toUpdate.Dequeue())(__instance);
             }
 
@@ -379,7 +380,7 @@ namespace sharp_injector.Helpers {
 
         public static void UpdateShortcutText(Window predictionWindow, int idx, SortedSet<Keys> keyboardShortcut) {
 
-            if(predictionWindow is null) {
+            if (predictionWindow is null) {
                 Terminal.Print("predictionWindow is null\n");
                 // Add this with the ability to update the prediction window.
                 toUpdate.Enqueue((pw) => UpdateShortcutText(pw, idx, keyboardShortcut));
@@ -389,7 +390,6 @@ namespace sharp_injector.Helpers {
                     var numbersSP = (StackPanel)pwType.GetField("PredictionNumbers", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static).GetValue(predictionWindow);
                     var border = (Border)numbersSP.Children[idx];
                     if (keyboardShortcut is null) {
-                        Terminal.Print("keyboardShortcut is null\n");
                         border.ToolTip = null;
                         ((TextBlock)border.Child).Text = "";
                         return;